+++ docs/source/conf.py
@@ -20,7 +20,7 @@ author = "Matthew Love"
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 nitpicky = True
-nitpick_ignore = [('py:data', 'typing.Union')]
+nitpick_ignore = [("py:data", "typing.Union")]
 
 extensions = [
     "sphinx.ext.autodoc",  # Generate docs from docstrings
@@ -29,7 +29,7 @@ extensions = [
     "sphinx.ext.intersphinx",  # Link to other projects' docs
     "sphinx.ext.viewcode",  # Add links to source code
     "sphinx.ext.githubpages",  # Auto-generate .nojekyll for GH Pages
-    "sphinx_argparse_cli", # argparse
+    "sphinx_argparse_cli",  # argparse
     "myst_parser",  # Parse Markdown files
     "module_table",  # Generate module catalog table
 ]
@@ -57,9 +57,7 @@ html_theme = "pydata_sphinx_theme"
 # html_static_path = ["_static"]
 
 # html_sidebars = {"**": []}
-html_sidebars = {
-    "modules/*": []
-}
+html_sidebars = {"modules/*": []}
 
 html_theme_options = {
     "github_url": "https://github.com/continuous-dems/fetchez",